rename topics middleware file to model and move it to the model folder, also change test so I check each item in the response
